Extract SettingsDetailRow
